refactor: SettingsDialog (687行) をUIコンポーネントとして分離
設定モーダル全体を独立コンポーネント化:
- チクタク音/アラーム/通知/振動パターン設定
- デフォルト値設定（ミーティング/ポモドーロ）
- アカウント・クラウド同期設定

comm-time.tsx: 3288行 → 2629行 (-659行)
累計: 5326行 → 2629行 (-51%)
